fix(SFT-1310, #1412): refactor exception catching, add refresh token check flow on `401` responses
### Related Ticket Number
[SFT-1310](https://solspace.atlassian.net/browse/SFT-1310)
#1412

### Description
A client reports that his ZOHO integration returns an `invalid_token` error, and based on the logs it seems it happens every hour, which is the lifespan of the `access_token`. For some reason his token does not get refreshed in a timely manner. A failsafe for this is added in this PR, which performs a token refresh flow once if an access token returns an unauthorized response.

- refactor failed response catching flow
- add a logger on failed exceptions through an event, rather than manually
- attach token refresh flow on `401` responses on refresh token capable integrations


[SFT-1310]: https://solspace.atlassian.net/browse/SFT-1310?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ